Fixed bug with incorrect countdown being shown.
Occurs when a user types in the URL of a page that has a flip on it and that URL is one that has already been visited before. Chrome (and possibly other) browsers pre-render the page as part of their URL suggestion logic. This background pre-rendering would cause flip to immediately sleep() before calculating tickExpectedTime.